requirements.txt
numpy
scipy
pyyaml
matplotlib
imageio